Implement modeling of asynchronous security events.
The ability to handle asynchronous events to the TSEM driver was
implemented.

In order to take advantage of this support was added for
injecting the asynchronous events into a security model by the
userspace trust orchestrator.

Initial support was added for shutting down the workload if an
out of model event was detected.  This is required since the
process will have moved forward past the security event and the
only option is to shutdown the workload.