Properly handle enforcing mode for atomic context security events.
Testing indicated that the trust orchestrator for an externally
modeled security namespace was always killing the workload if a
security event running in atomic context violated the model for
the namespace.

A statically scoped boolean variable named Enforce was added that
is set if the -e command-line argument is specified to the trust
orchestrator.  The value of this variable is referenced to
determine whether or not the workload should be killed if a model
violation occurs.

If the value of this variable is false the workload is not killed
and the violating event is registered as a forensic violation for
the model.