Updates to TSEM kernel documentation.
In preparation for the second release of TSEM for inclusion in
the mainstream kernel the TSEM documentation is being updated to
address issues that were raised by the Linux security maintainer.

This commit represents the first set of these changes.

A nomenclature is now being followed where the points in a model
are not being referred to as 'points' but exclusively as security
state coefficients or simply coefficients.

Previously a 'modeling domain' had been used as a description for
the TSEM namespaces.  Domain was criticized by Casey Schaufler in
another LSM as too generic.  Moving forward the TSEM security
namespace will be referred to as a security modeling namespace.