Update utilities to access new directory layout.
In the interest of future proofing the implementation the
directory that holds the pseudo-files for accessing the internal
trusted modeling agent was changed from the following location:

/sys/kernel/security/tsem

To the following directory:

/sys/kernel/security/tsem/InternalTMA/model0

This will allow alternate modeling implementations to be surfaced
by creating directories in the InternalTMA directory.

This commit modifies the trust orchestrators to use this
alternate location.

Work was also introduced to change nomenclature from referring to
security state coefficients to 'coefficients' rather than
'points'.