Improve error handling in trust orchestrator code.
Testing indicated that the trust orchestrators for both
internally and externally modeled domains was not behaving well
in the face of an error condition in the trust orchestrator code.

The behavior that was most problematic was the launch process
dying and leaving the monitor and/or workload processes still
running.  In the case of a container based workload this
typically left runc running in a detached state and interfering
with the console I/O that had been going to the launch process.

This problem was fixed by checking the return code from the child
monitor by the launch process and then having the launch process
shutdown the workload.  In the case of a container based workload
this is done by running the runc utility in 'kill' mode.  In the
case of a process workload a SIGKILL is sent to the lead process
in the workload.

The userspace trust orchestrator was the simplest case to handle
since the process monitoring the child was running with CAP_ML
privileges.  This allowed that process to send a termination
signal to the workload process.

In the case of the in kernel orchestrator the pipe that is used
to return the trajectory or security state coefficients was
second purposed to return the PID of the workload process as the
first line of output in the pipe.  This provides the launch
process, that has sufficient privileges, to terminate the
workload process.