Functional update to Xen based trust orchestrator.
Testing of the new TSEM driver had been validated with the
internal TMA and with the userspace based trust orchestrator.

This commit updates the SanchoXen implementation and the Xen
trust orchestrator to work with the new TSEM kernel interface and
to output the trajectory coefficient counts.

The trust orchestrator was also updated to handle orchestrator
failures more gracefully as was done with the userspace
orchestrator.