Support separate trajectory and forensics parameters.
This commit updates the userspace and Xen TMA orchestrators to
handle the output of separate coefficient and count values for
the execution and forensics trajectories.

This allows the following model query commands to be issued:

show trajectory
show counts
show coefficients
show forensics
show forensics_coefficients
show forensics_counts

For each orchestrator the routines to send the coefficients and
counts were re-purposed to have a single function do the output.
This was implemented by passing a flag variable that indicates
the type of each value to be output, either valid for an
execution trajectory or invalid for a forensics trajectory.

This work illuminated a somewhat subtle issue with the scheme
used to process command requests.  In the case of the forensics
commands the 'show forensics' syntax and command argument needs
to be after the 'show forensics_counts' and 'show
forensics_coefficients' arguments.  If the 'show forensics'
command is first in the last the scan of the command will
terminate with that command rather than continueing to find the
full command.