Compilation fixes.
In order to establish a full and complete build changes were made
to the command interpreters for the NRF52840 Sancho
implementation in order to change the command syntax for managing
security state coefficients from points to coefficients.

Additional work needs to be done on this TMA and its orchestrator
in order to support the new TSEM kernel interface that separates
standard trajectory coefficients from the forensics coefficients.