Eliminate use of strcpy in Cell object.
The standard C library supplied for the SGX development
environment does not surface the strcpy() function, presumably
for security reasons.  It appears their preference is for the
strcpy_s() function that is not surfaced by other standard
library implementations that are needed to support the full set
of implemented orchestrators.

Convert the strcpy implementation to equivalent functionality
using memcpy.