Update SanchoSGX implementation.
This commit brings the SanchoSGX TMA implementation up to date
with the new TSEM kernel control plane and infrastructure.  This
includes the general functionality updates that have been
previously documented for the userspace based trust orchestrator.

A flag was added to the ecall1 interface structure that allows
the synchronous vs asynchronous context of the event to be passed
into the enclave.  The SGX TMA implementation will not attempt to
issue an OCALL to set the trust status for an event that has been
generated and will simply return to the caller with the trust
status.  The trust orchestrator will terminate a workload that
generates an out-of-model event while in enforcing mode.

The get_point() SGX function was updated to accept a pointer to
an unsigned 64-bit in order to return the occupancy count for the
point.

The SanchoSGX->load_enclave and the SanchoSGX->load_enclave_memory
methods were extended to take a TSEMcontrol object that will be
used to issue the OCALL control plane calls.

The SecurityPoint object was updated to implement the following
two new methods:

SecurityPoint->set_count()

SecurityPoint->reset()

These methods are used by the SGX trust orchestrators to create
SecurityPoints based on the ECALL return values from the
get_point() function.