Use key=value arguments for trust control commands.
The TSEM driver is being updated to use key=value pairs for any
TSEM control commands that take an argument.

This is in response to the Linux LSM maintainer indicating that
API's need to be stable for 20+ years.  Absent using a version
number in the API, the only tractable way to provide that type of
extensibility is to use key=value pairs for arguments.

This eliminates the need to depend on positional parameters since
the kernel can always locate an argument that it is looking for.

This commit updates the 'trusted' and 'untrusted' control
commands to be of the following form:

trusted pid=PID key=HEXID

untrusted pid=PID key=HEXID

Where PID is the ASCII base10 encoded value for the PID that is
the subject of the control command.

HEXID is the ASCII base16 encoding of the authentication key that
the trust orchestrator had configured the security modeling
namespace to use when the namespace was created.