Fix display of forensics trajectory.
When the trust orchestrator management commands were updated to
match the new TSEM model0 interface definitions the 'forensics'
command was placed after the 'forensics_counts' and
'forensics_coefficients' command in order to block matching on
the 'forensics' portion of the latter two keywords.

Unfortunately that work had neglected to change order of the
enumerations in order to match the command order.  This was
causing the forensics command to generate output forensics_counts
output when the quixote-console utility was run in one-shot mode
with the -F argument.

This was corrected by verify that the enumeration types matched
the array index locations of the command syntax that invoked
these commands.