Address segmentation fault in pseudonym handling.
Testing of the userspace orchestrator indicating that a
segmentation fault was being generated on the processing of the
first security event in a workload, when a pseudonym was defined
in the model.

This was traced to the handling of an asynchronous event,
although the nature of the regression indicated it should have
happened with any event.

The root cause is that the String object that contains the
pathname was returning a null pointer since a filename had not
yet been committed to the object.

The fix was to simply check for the return of a null pointer and
terminate pseudonym processing.  Since the Buffer object that
holds the pseudonym is reset there is a guarantee that the event
will be treated as the absence of a pseudonym.