Fix pseudonym generation.
When the event export had been changed to export a pathname
rather than the digest of the filename the pseudonym handling had
not been updated.  It was adding the pathname of the file rather
than the digest of the filename.  This was causing an incorrect
pseudonym value to be generated that in turn failed to cause a
file to be recognized as a pseudonym.

Correct this by hashing the filename as part of the generation of
the pseudonym value to be compared against the pseudonym list.