Automate source distribution creation.
The 'git archive' command does not generate files that are in a
submodules directory so the HurdLib support directory is not
included in the distribution.

In addition the primary GIT repository has a number of
directories that are not for public distribution.

Create a 'tar' target for the primary Makefile that creates a
directory containing the desired source distribution and then
generates a tar archive from that directory.

Introduce configuration variables in the Config.mk file that
allows a release, version name and canonical name for the
package.