Expose task disambiguator value.
The TSEM driver was enhanced to output a task temporal
disambiguator value.  This value is used to identify separate
instances of a task identity.

At the current time the modeling engine does not treat this
value, the support is primarily to allow the value to be
presented in the trajectory output.