Surface event timestamp information.
The TSEM drive was updated to include the ability to capture the
time in nanoseconds since boot when the event occurred.  The
clock source selected for the time was a block that includes any
time that was spent in suspend.

This timestamp value is included as the ts: member of the event
structure.  Update the TSEMevent object to expose this value in
trajectory outputs.