Drop concept of generically modeled events.
The TSEM driver was updated to no longer use the notion of a
'generic_event' type.  Instead, all events are treated as first
class events with the previously generically modeled events
having a null set of CELL parameters that is denoted by the JSON
null field descriptor, {}.

Translation was straight forward and the logic in Cell object is
to explicitly denote the security events that are parameterized
and use the default: clause of a switch statement keying on the
security event type desdcriptor to handle a generically modeled
event.