Support export only security modeling namespaces.
The TSEM driver was enhanced to support the concept of 'export
only' security modeling namespaces.  These namespaces simply
export the security event descriptions and do not wait for the
external TMA to release the process with an event status.

An alternate model would be to simply have an external TMA always
signal success for a security event but this would impose the
need for the process to sleep while the TMA conducts the
approval.  The export only namespaces imply only the amount of
overhead needed to setup the security event description export
and submission to the export workqueue.

The following utility was created to provide a minimal
implementation of the infrastructure needed to capture an event
stream:

quixote-export

It can run in either container mode or process mode.

The utility requires a -o argument to specify the output file
that the security event description stream is written to.

The TSEMcontrol object was updated to implement the following
additional control type:

TSEMcontrol_TYPE_EXPORT

Setting this event type for the call to the ->create_ns method
causes the 'export' keyword to be written to the TSEM control
plane, rather than external or internal.  Any of the additional
characteristics for configuring a security modeling namespace are
available as well.