Implement quixote-cat utilility for export only of root events.
The kernel TSEM driver was updated to implement an 'export-only'
mode for the root security modeling namespace.  This mode is
selected by specifiying tsem_mode=2 as a kernel command-line
argument.

In this mode all of the security events that occur in the root
modeling namespace are surfaced through the following pseudo-file
in the TSEM control plane:

/sys/kernel/security/tsem/ExternalTMA/0

Security events are stored until they are read so the events must
be either continually read or read on a frequent basis in order
to avoid consuming all of the platform memory.

The quixote-cat utility was implemented to provide a utility for
reading the root security events.  This utility operates in one
of two modes; a simple 'cat' mode that reads all of the current
events that are storage and 'follow' mode where all of the
current events are read with the utility then continueing to read
and output events as they are generated.

Follow mode is selected by the -f command-line argument to the
utility.

Since the quixote-cat utility will be running in the same
security modeling namespace that it is processing the event
stream will contain artifcats from the reading and export of the
events.  A common method of capturing events is with the
following command:

quixote-cat | ncat HOSTNAME PORT

Which is a simple method of sending events to a remote host for
modeling.

A simple invocation as above will generate large numbers of
'socket_sendmsg' events.

The generation of these artifacts can be reduced by using the -b
command line option to specify a larger buffer size to
accumulate events before the are pushed to the standard out
stream.  The -b command-line option takes an integer argument
that is the number of megabytes of memory to be allocated for the
output buffer.

Using a one megabyte buffer will typically generate measurement
artifacts that consist of approximately five percent of the total
number of events that are generated.