Support export of event descriptions with lineage tracking.
The kernel TSEM driver was updated to export the parent task
identity and the parent task temporal disambiguator for each
security event.

This allows the tracking of the lineage of security events.  In
other words the security characteristics of the task that that
led to the task generating a security events.

This allows, for example, the differentiation of security events
that came from the base instantiation of a process or one that
occurred from a forked copy of a process.

A security event that occurs in the base (initial) instance of an
executable will have the task identity of the process that
generated the execution of the binary.  A forked copy of a
process that is conducting actions on behalf of the process will
have the task identity of the process that it was forked from.

The parent ttd uses a label of p_ttd while the parent task
identity uses a label of p_task_id.

This patch conveys the descriptions into the external TMA's but
the models do not currently incorporate the parent task identity
as a characteristic of the security event.