Handle attributes for inode_getattr.
The process of adding support for full modeling of all security
events was started in the TSEM driver with implementation of
parameters for the inode_getattr LSM hook.

This exposed the fact that the JSON->Quixote translation code
needs to be updated on an event by event basis.

This suggests that the best path forward is to expend the
resources to move to using JSON encoding in the security modeling
code rather than imposing translations to and from these two
different but related formats.

This commit provides support for the handling of the
inode_getattr event type in order to have the userspace
compatible with the current state of the kernel driver.