Transition to JSON based event processing.
This commit initiates the process of converting the Trusted
Modeling Agents to correspond in the native JSON export format
that the TSEM kernel driver uses.  This drops the need for the
external TMA's to indulge in converting the JSON export format to
the previous Quixote based formatting.

The trust orchestrators for internally modeled domains was
initially converted as was the orchestrator for userspace based
modeling agents.