Implement support for build of mosquitto MQTT broker and client.
The decision has been made to use the MQTT protocol in order to
support the export of security event descriptions.  The
mosquitto implementation is one of the most straight forward
implementations of both a client and broker.

Implement support for building a selected version of mosquitto in
the Support directory.