Implement export of root modeling events in quixote-export.
Previously, the quixote-cat tool had been implemented to export
the security event that were being generated in the root security
modeling namespace.

With the addition of MQTT export in the quixote-export tool it
made little sense to maintain a separate utility for exporting
root security events, when the additional functionality could be
easily added to the quixote-export utility.

The -R option was added to the quixote-export utility in order to
specify that modeling of the root security modeling namespace
should be conducted.

Initial testing of the utility indicated a significant regression
in that the quixote-export utility, when run in broker mode,
caused depletion of all kernel memory.  This appeared to be
secondary to the large number of network related security events
that were generated by forwarding of the MQTT messages.

Further testing indicated the issue could be resolved by queueing
up a number of security events before publishing them to the
broker.  The events were queued up with embedded linefeed
characters so subscriptions to the broker would appear to be the
same as if the messages were individually exported.

A default queue size of 100 events was selected.  The -q
command-line option was implemented to allow this to be
configured on an execution specific basis.

This functionality is similar to what had been implemented with
the -b command-line option for the quixote-cat utility which
increased the size of the standard output buffer in order to
decrease the impacts of self-measurement.

The default file output option was configured to be /dev/stdout,
so that invoking root export without specifying broker mode would
result in output being generated to the standard output stream.
The queuing function is also in effect when output is being
generated to a file so altering the queue size can be used to
minimize self-measurement artifacts when sending output to
standard output or a file.