Fix empty lines in trajectory export output.
Testing indicated that extra linefeeds were being injected into
the output from the quixote-export utility when output was
directed to a file or standard output.

The linefeeds were being generated at consistent locations in the
output.  After some examination the locations in the output
stream were consistent with the output queue size that was being
used.

This allowed the problem to be sourced to the outpu_event()
function.  The code was adding a linefeed to the contents of the
String that had been generated from the trajectory events.  This
extra linefeed was causing the empty lines to be inserted in the
output every N lines where N was the output queue size.

Remove the insertion of the linefeed in order to address this
issue.