Update parsing of file_open handler.
This commit begins the process of updating the TSEM event parser
to match the updated JSON syntax that is generated by kernels
that have full parameterization support.

This involves introducing structures such as the inode and path
structures that match such structures, at least in intent, that
are in the TSEM kernel implementation.

A new strategy will be followed where the handler for each CELL
type is a first class function rather than a helper function, ie.
prefixed by an underscore.