Restrict selection of keys to top level of current description.
A regression was noted where the first key that is found in a
JSON encoding is selected for value extraction.  The expectation
is that the TSEMparser library be used in a manner where only top
level keys in the current JSON encoding are extracted.

The API requires that subordinate structures be selected for
operation with desired keys being selected from those
sub-structures.

Introduce the _find_level() function to determine what nesting
level the key that has been currently found is at.  The search
for a key only returns if the nesting level is zero.

A loop was introduced to the _get_key() function that scans
across the current JSON description looking for a level 0 key
that matches the key whose value has been selected.