Update mmap_file CELL parsing.
This commit modifies the parsing of a TSEM_MMAP_FILE event
description so that is consistent with the new TSEM JSON encoding
format.

These modifications involved significant changes over the
entirety of the CELL object.

The _get_16() function was added to allow the extraction of a 16
bit integer value.  This is to handle the fact that the mode
variable of an inode is to be encoded as a 16 bit value.

The parsing a file{} structure was isolated into the
_parse_file() function so that it could be used by both the
TSEM_MMAP_FILE and TSEM_FILE_OPEN events.  The initiation of the
parsing of a TSEM_FILE_OPEN event is now done by the
parse_file_open() function.

Following a similar paradigm the formatting of a file{}
description was isolated into the _format_file() function.  The
_format_file_open() function was introduced as the the method of
initiating the formatting of the TSEM_FILE_OPEN event.