Feed the correct file definition stream into the CELL mapping.
The hash of the file coefficient stream was being fed into the
input of the hash function that will generate the CELL
measurement rather than the the coefficient stream itself.

Drop the use of the Sha256 object in the _measure_file() and
simply feed the buffer contents into the hash function that will
generate the CELL mapping.