Implement support for new socket descriptions.
This commit updates the Cell parsing logic to handle the new JSON
formatting of security event descriptions for socket security
operations.

The sock struct was introduced to encapsulate basic socket
information including the task identity.  This structure was
incorporated into the socket_connect_parameters and the
socket_accept_parameters structures.

The _parse_sock() function was introduced to handle the parsing
of the JSON sock {} description into the sock structure.

The functions that drive the parsing of the socket events that
are currently handled were capitalized, following the scheme that
was adopted for the file handling events.

The _measure_sock() function was introduced to generate the
security state coefficient stream based on the sock structure.

The _format_sock() function was introduced to handle the
formating of the information in a sock structure back into a JSON
sock {} description.