Use CAP_MAC_ADMIN for security namespace access control.
The TSEM driver was updated to drop the use of a new security
capability bit (CAP_ML) in favor of CAP_MAC_ADMIN.  This was
driven by Linux maintainer resistance the implementation of a new
capability bit for the purpose of controlling security modeling
namespaces.

From a security perspective this is an inferior implementation
given that the ability to sequester a process heirarchy into its
own modeling namespace is a significant method of evading model
based security controls and the notion of a trust orchestration
environment.

It also precludes the ability to use label based mandatory access
controls that require administrative capabilities inside of a
security modeling namespace, given the fact that trust
orchestrators drop their administrative capability bit when
starting the workload.  One of the uses for such a model would be
to verify the type and value of extended attributes used to hold
security labels.