Increase and set common size for TSEM event descriptions.
A regression was noted with the userspace trust orchestrator in
the face of the fully parameterized security event descriptions.
The TSEM_INODE_RENAME event was generating an event description
larger than the buffer that was allocated for the read of the
event description.

This buffer needs to be converted to dynamic allocation but until
then increase the buffer size to a comfortable margin over the
largest observed event description.

This commit standardizes the definition for that buffer size in
the quixote.h header file so that it can be easily changed in the
future.