Use null UUID for inodes on a tmpfs filesystem.
Since the UUID value for a tmpfs based filesystem changes with
every mount it is not possible to have an invariant state
function for inodes on these filesystems.

Use the same approximation that the in-kernel TMA uses to address
this issue by using a null UUID (16 null byte) as the measurement
value for an inode based on a tmpfs filesystem.