Remove support for TSEM specific capability.
The version 3 release of TSEM dropped the notion of using a new
capability bit to arbitrate the creation of a security modeling
namespace.

This met with significant resistance from the Linux security
maintainers who advocated that the existing CAP_MAC_ADMIN
capability should be used.  This restricts the ability to use
existing LSM's in the context of a security modeling namespace
since the Quixote utilities drop CAP_MAC_ADMIN before starting
the workload.

Since forward progress cannot be made with respect to mainstream
acceptance of TSEM this commit drops support for building a
modified version of libcap.