Conditionalize build of SGX components.
Disabling the build of the SRDE directory resulted in errors when
the orchestration utilities were being built since the components
needed for the quixote-sgx and quixote-sgx-u utilities were not
being built.

Conditionize the build of the SGX infrastructure on the presence
of the SGX member of the BUILD_SANCHOS directive.