Build Support library in proper order.
A build from a pristine source directory was failing secondary to
the fact that the lib directory was getting built before the
Support directory.  The Support directory contains necessary
pre-requisites for the lib directory.

Update the Makefile so that the Support directory is treated like
the HurdLib directory rather than as part of the directories
specified by the SUBDIRS variable.