Use pkg-config definitions for libcrypto and libssl.
The build of the MQTT infrastructure was failing on GCC secondary
to the necessary OpenSSL libraries not being specified.

Address this by using the pkg-config command to extract the
library definitions for both the libssl and libcrypto libraries.

The real time library is needed with the GCC implementation, a
fact that is not recognized by the pkg-config system.  Specify
the -lrt library which does not impact the musl-gcc build.