Fix build infrastructure for mbedtls.
A regression was noted when building Quixote from a native tar
based source distribution.  Testing indicated that it appears as
if Github had changed its conventions for downloading a tar
archive.

This was corrected along with a change to force a timestamp of
the final source directory in order to prevent a second
invocation of a build from unpacking a second copy of the mbedtls
archive into the mbedtls directory that was created by the
initial generation of the directory.