Create Build directory for support files.
The build infrastructure in the Support directory places the
downloaded source files in the Build directory but this may not
be present in a fresh source archive.

Modify the Makefile in order to create the directory if it does
not exist.  The purge directive removes this directory to return
the tree to a pristine state.