Fix export modeling.
Issues were noted where export only logging to a file was not
including linefeeds at the end of each event description.

The needs of logging to a file and to an MQTT broker are
different, as the MQTT output does not need linefeed termination.

Introduce the output_queued() function in order to handl the
generation of output that has been accumulated in the event
queue.  By default queued functions have linefeeds added to them
so strip the linefeeds before submitting the strings to the MQTT
broker.

The incorrect filename was being used for the export file,
secondary to the filename location having changed in the TSEM
sources.  Correct the define to use the correct pathname for the
file.