Fix regressions in quixote-export orchestrator.
Testing of event capture indicated a number of problems that this
commit addresses.

Even if the -b argument was set to place the event orchestrator
in broker mode output was generated to standard output.  This was
corrected by only opening and initializing the File object if the
broken argument had not been specified.

File based I/O did not have linefeed terminations after the end
of each event.  Broker based output had an extra linefeed at very
N records where N was the queue size in case of root export
modeling.

The output_event_description() function was added whose only role
was output the contents of the Output_String variable to the
requested output file.

The output_queued() function was added that is responsible for
taking actions necessary to output a set of queued event
descriptions.  In this made a carriage return is not added for
file based output since every record is linefeed terminated.

The output_event() function is called when events are set on an
event by event basis.  This function appends a carriage return
to descriptions that are being output to a file.