Implement support for specifying a loaded security model.
With the TSEM kernel driver gaining support for security models
implemented with loadable models the TSEMcontrol object was
updated to allow the name of a model to be requested.

The fourth argument of the TSEMcontrol->create_ns method was
converted to be name of the model that is to be requested.

If this argument is non-null the following option is added the
message that is written to the TSEM control plane file:

model=MODELNAME

Where MODELNAME will be replaced with the string pointed to by
the fourth argument of the TSEMcontrol->create method.

The trust orchestrators were updated to accept the -M
command-line argument that is used to specify the name of the
model that is to be requested.