Update root export implement for quixote-export.
Testing the export of security events in an OS instance running
with root namespace only modeling indicated that the exported
events were not being cleared from the event list running in the
event structures depleting available system memory.

The TSEM kernel implementation appeared to be correct so
attention was directed at the quixote-export utility.  It appears
that there may have been some type of issue with the reading of
events in 'follow' mode that was not causing them to be cleared
from the event export list.

The previous implementation simply appended events to a string
that was maintained in the TSEMevent object.  This implementation
had a number of other issues that were problematic.

The export of root events was completely re-written to take
advantage of a Gaggle to store all of the events until the event
queue limit was reached.  This resulted in a much cleaner
implementation that will have benefits in the export of events
from a subordinate security modeling namespace.

The local GWHACK implementation, that should not have been
maintained in this utility was eliminated.  The Gaggle object is
populated with a set of String objects up to the size of the
event queue that is specified.

A new statically scoped variable named Queued was implemented to
track the number of events that have been set in the Gaggle.
This variable is reset each time the set of queued events is
flushed.

The following helper functions for the export_root() function
were added:

_output_events()

_export_events()

The _export_events() function is responsible for reading, queuing
and flushing events until an end of data condition is met.

The _output_events() function is called when the queue is filled
to generate the output to either the specified output file or to
an MQTT broker.

The previous output_queued() function is no longer needed and was
eliminated.

All of the events generated through root modeling have linefeed
terminations on them.  The MQTT subscription client should thus
not add linefeeds as that would result in an extraneous linefeed
at the end of each queued block of events.