Linefeed terminate export messages from modeling namespaces.
In order to have consistency between the export of messages from
the root modeling namespace and those from subordinate modeling
namespaces all messages from subordinate modeling namespaces are
linefeed terminated with this commit.

The output_event_description() function was eliminated given that
its functionality was folded into the output_event() function.

Since the export messages processed by the TSEMevent method have
their linefeeds removed a linefeed is added in the output_event()
function so that both MQTT as well as file based messages are
linefeed terminated.

If the mosquitto_sub utility is used to subscribe to messages
from a broker the -N command-line argument needs to be used so
that the client does not add a redundant linefeed.