Use common lseek after export file read.
After the export pseudo-file was successfully read a lseek call
was issued in order to prepare for the read of the next record.

The code that was reading the export pseudo-file in a subordinate
modeling namespace was also issueing a lseek after a read
returned with a NODATA indication.

Use a common lseek in the the read of the root export pseudo-file
in order to have common behavior between both venues.