Update HurdLib for Process object support.
The Process object was added to HurdLib in order to simplify the
execution of an executable task.

With respect to the trust orchestrators the Process object has
the ability to execute a command that is delimited from other
command-line arguments by a -- character sequence.

This functionality will be used to implement the -X mode of the
trust orchestrators that will be used to specify a command that
is to be run in a subordinate modeling namespace.