Add -X mode support for quixote-export utility.
This commit implements a new operating mode for the
quixote-export trust orchestrator.  The -X mode is used to
directly execute a command that is specified with arguments on
the quixote-export command-line.

The command to be executed is delimited from the quixote-export
command options with the -- character delimiter sequence.

For example the following quixote-export command:

quixote-export -X -o export.json -- wc -l /etc/passwd

Will execute the following command:

wc -l /etc/passwd

In a subordinate modeling namespace and write the execution
trajectory into the export.json file.