Add MQTTduct->set_password method.
In order to provide more flexible invocation options for the
quixote-export trust orchestrator the ->set_password method was
added to the MQTTduct object to allow more flexible configuration
of the MQTT authentication password.

A String object was added to the MQTTduct state object to hold
the password that is to be used to authenticate to the broker.
The password to be used is added to this object by calling the
->set_password method.

The ability to set the password directly by the ->init_publisher
method is retained.  The password that is configured by the
->set_password method is selected by passing a NULL value as the
password argument to the ->init_publisher method.  Setting a
password with the ->init_publisher argument causes the stored
password value to be overridden.

The ->set_password object first checks for the following
environment variable:

TSEM_PASSWORD

If this environment variable is set it is saved as the default
password.

If this environment variable is not set the contents of the
following file is used as the source of the password:

/opt/Quixote/etc/mqtt.pwd

If both methods fail to set a default password the ->set_password
call fails.