Use MQTTduct->set_password to set MQTT broker password.
This update modifies the quixote-export utility to use the newly
introduced ->set_password method to set the password that will be
used for subscribing to the MQTT broker.

The primary motivation for this change is to take advantage of
the ability to define the MQTT broker password in a file rather
than using an environment variable.  This in turn allows
individual applications to be launched in their own modeling
namespaces by init or systemd.